chore: move StorageRootHash out of zk module
