+++ crates/sovereign-sdk/rollup-interface/src/state_machine/zk/light_client_proof/output.rs
@@ -2,7 +2,7 @@ use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize};
 
 use crate::da::LatestDaState;
-use crate::zk::StorageRootHash;
+use crate::StorageRootHash;
 
 /// The output of light client proof
 #[derive(Debug, Clone, BorshDeserialize, BorshSerialize, PartialEq)]
